requirements.txt
parsy
attrs
mypy
flake8